model-inference/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pandas==1.5.3
joblib==1.2.0
numpy==1.23.5
scikit-learn==1.2.2
category_encoders==2.5.1
python-dotenv==1.0.0